+++ ichnaea/async/schedule.py
@@ -23,66 +23,6 @@ CELERYBEAT_SCHEDULE = {
         'options': {'expires': 570},
     },
 
-    # BBB: Continuous location update tasks
-
-    'monitor-observations': {
-        'task': 'ichnaea.monitor.tasks.monitor_measures',
-        'schedule': timedelta(seconds=900),
-        'options': {'expires': 990},
-    },
-
-    'location-update-cell-1': {
-        'task': 'ichnaea.data.tasks.location_update_cell',
-        'schedule': timedelta(seconds=29),
-        'args': (1, 1000000, 1000),
-        'options': {'expires': 25},
-    },
-    'location-update-wifi-1': {
-        'task': 'ichnaea.data.tasks.location_update_wifi',
-        'schedule': timedelta(seconds=31),
-        'args': (1, 1000000, 1000),
-        'options': {'expires': 25},
-    },
-
-    # BBB: Daily backup tasks
-
-    's3-schedule-cellobservation-archival': {
-        'task': 'ichnaea.backup.tasks.schedule_cellmeasure_archival',
-        'args': (1000, 100000),
-        'schedule': crontab(hour=1, minute=7),
-        'options': {'expires': 43200},
-    },
-    's3-write-cellbackups': {
-        'task': 'ichnaea.backup.tasks.write_cellmeasure_s3_backups',
-        'args': (1000, 10000, 60),
-        'schedule': crontab(hour=2, minute=7),
-        'options': {'expires': 43200},
-    },
-    's3-schedule-wifiobservations-archival': {
-        'task': 'ichnaea.backup.tasks.schedule_wifimeasure_archival',
-        'args': (1000, 100000),
-        'schedule': crontab(hour=1, minute=17),
-        'options': {'expires': 43200},
-    },
-    's3-write-wifibackups': {
-        'task': 'ichnaea.backup.tasks.write_wifimeasure_s3_backups',
-        'args': (1000, 10000, 60),
-        'schedule': crontab(hour=2, minute=17),
-        'options': {'expires': 43200},
-    },
-    's3-delete-wifiobservations': {
-        'task': 'ichnaea.backup.tasks.delete_wifimeasure_records',
-        'args': (1000, 1, 120, 1000),
-        'schedule': crontab(hour=3, minute=17),
-        'options': {'expires': 43200},
-    },
-    's3-delete-cellobservations': {
-        'task': 'ichnaea.backup.tasks.delete_cellmeasure_records',
-        'args': (1000, 1, 120, 1000),
-        'schedule': crontab(hour=3, minute=27),
-        'options': {'expires': 43200},
-    },
-
     # Hourly
 
     'update-statcounter': {