+++ ichnaea/models/__init__.py
@@ -2,11 +2,6 @@ from ichnaea.models.base import _Model
 
 # import all models, to make sure they are all registered
 from ichnaea.models.api import ApiKey  # NOQA
-from ichnaea.models.backup import (  # NOQA
-    ObservationBlock,
-    ObservationType,
-    OBSERVATION_TYPE_META,
-)
 from ichnaea.models.cell import (  # NOQA
     Cell,
     CellArea,