+++ ichnaea/tests/factories.py
@@ -1,4 +1,3 @@
-from hashlib import sha1
 from inspect import ismethod
 
 import factory
@@ -14,8 +13,6 @@ from ichnaea.models import (
     Cell,
     CellArea,
     CellObservation,
-    ObservationBlock,
-    ObservationType,
     OCIDCell,
     OCIDCellArea,
     Radio,
@@ -101,19 +98,6 @@ class CellObservationFactory(CellPositionFactory):
 
     class Meta:
         model = CellObservation.create
-
-
-class ObservationBlockFactory(BaseFactory):
-
-    class Meta:
-        model = ObservationBlock
-
-    measure_type = ObservationType.cell
-    start_id = 10
-    end_id = 20
-    s3_key = '201502/cell_10_200.zip'
-    archive_sha = sha1('').digest()
-    archive_date = None
 
 
 class WifiPositionFactory(BaseFactory):